Move OpenGarageConfigEntry to coordinator module
