+++ homeassistant/components/opengarage/coordinator.py
@@ -18,15 +18,18 @@ from .const import DOMAIN
 _LOGGER = logging.getLogger(__name__)
 
 
+type OpenGarageConfigEntry = ConfigEntry[OpenGarageDataUpdateCoordinator]
+
+
 class OpenGarageDataUpdateCoordinator(DataUpdateCoordinator[dict[str, Any]]):
     """Class to manage fetching Opengarage data."""
 
-    config_entry: ConfigEntry
+    config_entry: OpenGarageConfigEntry
 
     def __init__(
         self,
         hass: HomeAssistant,
-        config_entry: ConfigEntry,
+        config_entry: OpenGarageConfigEntry,
         open_garage_connection: opengarage.OpenGarage,
     ) -> None:
         """Initialize global Opengarage data updater."""
